Created Card Calculation System for BlackJack
Created an array that gets a random value calculated from 1-13 into it every time a fly is hit. Afterwards, a function creates and prints a string of those numbers on top of the screen. If the number exceeds 21, the values in the array get erased. If more than 5 cards are drawn, the "5-Card-Charlie" rule comes into effect meaning you win. For now, this also erases the string. Working on the fly's positioning next, then the gambling and the dealer.