Dealer and win/lose conditions created
Created a Dealer to play against. Calculations that determine win/lose/push conditions created, as well as game states for each outcome. Created a button to "stand", which determines many of these outcomes.

Next, I want to work on hiding the first card in the dealer's hand, and fleshing out the game states a little. Afterwards I want to start and finish the betting system. Lastly, a fresh coat of paint (with sound design) Will be added.