'Card in the hole' functionality and betting system
Created a way for the first card the dealer draws to be separate and hidden from the user. This also creates 2 sums, one of the cards faceup, and one of the full sum (that gets revealed when the card in the hole is also revealed). Also created a simple "gambling" system, with 3:2 odds. Also updated the movement of the flies to follow a circular pattern rather than a straight line.

The last few things I need to do are to create a few buttons, fix 2 bugs to do with the dealer, and set a maximum and minimum limit on what you can bet.  I may also add a bit more to the movement of the flies if I have the time.

Otherwise, besides those few things, I just have to add a coat of paint and I'll be done.