Fly and target movement, difficulty scaling
Created fly and target movement based on circles. Both of these objects bounce off the side of the screen, and the flies follow randomized circular movement.

Fixed maximum and minimum betting. Also added "difficulty scaling" where the speed of the flies and targets speed up marginally according to the amount of money you have in the bank. Fixed a few bugs as well, mostly involving the dealer.